full walkthrough of scala backend complete
