+++ scala-backend/app/saladbar/src/main/scala/ExprToStringError.scala
@@ -3,6 +3,8 @@ package saladbar
 
 /**
   * ExprToStringError
+  * 
+  * Used by the Parser class
   *
   * @param msg
   */
