+++ scala-backend/app/saladbar/src/main/scala/IfThenElse.scala
@@ -3,11 +3,8 @@ package saladbar
 
 /**
   * IfThenElse(e1, e2, e3)
-  * from concrete syntax: if (e1) e2 else e3
   * 
-  * if e1 evaluates to a truethy value,
-  *     then evaluate e2
-  *     else evaluate e3
+  * from concrete syntax: if (e1) e2 else e3
   */
 case class IfThenElse(e1: Expr, e2: Expr, e3: Expr) extends Expr {
 
@@ -15,6 +12,17 @@ case class IfThenElse(e1: Expr, e2: Expr, e3: Expr) extends Expr {
     override def toString: String = s"if ($e1) { $e2 } else { $e3 }"
 
     
+    /**
+      * step
+      *
+      * if e1 evaluates to a truethy value,
+      *     then evaluate e2
+      *     else evaluate e3
+      * 
+      * @param evalConditions
+      * @param sc
+      * @return
+      */
     def step[A](evalConditions: EvalConditions)(sc: Expr => A): A = {
         if (e1.isValue) {
             evalConditions.getTc.checkIf(e1.asInstanceOf[Value]){
@@ -30,6 +38,17 @@ case class IfThenElse(e1: Expr, e2: Expr, e3: Expr) extends Expr {
     }
 
 
+    /**
+      * substitute
+      * 
+      * find all free instances of $x in $this and replace with $esub
+      *
+      * @param evalConditions
+      * @param x
+      * @param esub
+      * @param sc
+      * @return
+      */
     def substitute[A](evalConditions: EvalConditions, x: String, esub: Expr)(sc: Expr => A): A = {
         e1.substitute(evalConditions, x, esub){
             e1p => e2.substitute(evalConditions, x, esub){
