+++ scala-backend/app/saladbar/src/main/scala/ImplicitConversions.scala
@@ -3,6 +3,13 @@ package saladbar
 
 /**
   * ImplicitConversions
+  * 
+  * Think javascript type changes, so 
+  * true + 5 == 6
+  * true && 5 == true
+  * "hello" + 5 == "hello5"
+  * "5" + 5 == "55"
+  * ... and more ...
   */
 case object ImplicitConversions extends TypeCondition {
 
@@ -11,7 +18,9 @@ case object ImplicitConversions extends TypeCondition {
     /**
       * performUop
       * 
-      * always convert to expected type of the logical operation
+      * always convert to expected type of the relevant operation
+      *   `!` i.e. `Not`: expects a logical operand
+      *   the others expect a numerical operand
       *
       * @param uop
       * @param v1
@@ -33,7 +42,7 @@ case object ImplicitConversions extends TypeCondition {
     /**
       * checkBop1
       * 
-      * always valid here.
+      * always valid here. We will type convert on the fly
       *
       * @param bop
       * @param v1
@@ -49,7 +58,7 @@ case object ImplicitConversions extends TypeCondition {
       * performShortCircuitBop
       * 
       * currently only And/Or support shortcircuiting ops
-      * here, alway convert v1 to true/false
+      * They each expect logical operands.
       *
       * @param bop
       * @param v1
@@ -69,8 +78,12 @@ case object ImplicitConversions extends TypeCondition {
     /**
       * performBop
       * 
-      * always convert to expected type of the logical operation
-      * for plus: if either is a string, convert both to strings.
+      * always convert to expected type of the relevant operation
+      * === and !== compare both type and value
+      * Comparison, == and != works on strings iff both operands are already strings,
+      *     else convert to numbers, then compare
+      * Plus: if either is a string, convert both to strings.
+      * remainder expects numbers.
       *
       * @param bop
       * @param v1
@@ -86,14 +99,9 @@ case object ImplicitConversions extends TypeCondition {
                 case _ => sc(B(f(v1.toNum, v2.toNum)))
             }
         }
+
         // TODO: bop class can have an (A, A) => B method
         bop match {
-            case Plus => (v1, v2) match {
-                case (S(_), _) | (_, S(_)) => sc(S(v1.toString + v2.toString))
-                case _ => sc(N(v1.toNum + v2.toNum))
-            }
-            case Times => sc(N(v1.toNum * v2.toNum))
-            case Minus => sc(N(v1.toNum - v2.toNum))
             case Geq => hCmp { _ >= _ }{ _ >= _ }
             case Gt => hCmp { _ > _ }{ _ > _ }
             case Leq => hCmp { _ <= _ }{ _ <= _ }
@@ -102,6 +110,17 @@ case object ImplicitConversions extends TypeCondition {
             case Neq => hCmp { _ != _ }{ _ != _ }
             case Eqq => sc(B(v1 == v2))
             case Neqq => sc(B(v1 != v2))
+            case Plus => (v1, v2) match {
+                case (S(_), _) | (_, S(_)) => sc(S(v1.toString + v2.toString))
+                case _ => sc(N(v1.toNum + v2.toNum))
+            }
+            case Times => sc(N(v1.toNum * v2.toNum))
+            case Minus => sc(N(v1.toNum - v2.toNum))
+            case Div => {
+                val n2 = v2.toNum
+                if (n2.isNaN()) sc(N(n2))
+                else sc(N(v1.toNum / n2))
+            }
         }
     }
 
@@ -109,7 +128,7 @@ case object ImplicitConversions extends TypeCondition {
     /**
       * checkIf
       * 
-      * with implicit conversion the value can always be made into true/false
+      * with implicit conversion the value can always be made into logical value true/false
       *
       * @param v1
       * @param sc
