+++ scala-backend/app/saladbar/src/main/scala/Interpreter.scala
@@ -1,29 +1,66 @@
 package saladbar
 
-case class InterpreterError(msg: String) extends Exception{
-  override def toString: String = s"INTERPRETER ERROR: $msg"
-}
 
 /**
-  * Template scoping conditions to effect substitution. sub out functions for closures.
-  * template type changes on value with hBop kind of things
-  * test all 4 options well.
+  * Interpreter
+  * 
+  * Given an expression and a set of evaluation conditions,
+  * Determine the sequence of steps during evaluation of the expression.
+  * 
+  * OO Pattern: Interpreter
+  * 
+  * @param evalConditions
   */
-
-
 class Interpreter(private var evalConditions: EvalConditions) {
 
+  /**
+    * setEvalConditions
+    * 
+    * allows you to change the evaluation conditions of interpretation
+    * if needed.
+    *
+    * @param ec
+    */
   def setEvalConditions(ec: EvalConditions) = 
       this.evalConditions = ec
 
+    
+  /**
+    * evaluate
+    * 
+    * limit users to the first 100 steps only
+    *
+    * @param e
+    * @param sc
+    * @return
+    */
   def evaluate[A](e: Expr)(sc: List[Expr] => A): A = evaluate(e, 100)(sc)
+
+  
+  /**
+    * evaluate
+    * 
+    * Attempt evaluation of $e limited by $maxSteps. If an error
+    * occurs in evaluation, let that be the terminating value of 
+    * evaluation captured as a `LettuceError` with the relevant
+    * error message captured.
+    * 
+    * e.g. for expression 1 + 2 * 3, sc is called on
+    *      List(1+2*3,
+    *           1+6,
+    *           7)
+    *
+    * @param e
+    * @param maxSteps
+    * @param sc
+    * @return
+    */
   private def evaluate[A](e: Expr, maxSteps: Int)(sc: List[Expr] => A): A = {
     if (maxSteps <= 0 || e.isValue) sc(List(e))
     else {
-      // TODO: capture lettuce error?
       e.stepWrapper(evalConditions){
-        ep => evaluate(ep, maxSteps - 1) {
-          t => sc(e :: t)
+        newE => evaluate(newE, maxSteps - 1) {
+          expressionSoFar => sc(e :: expressionSoFar)
         }
       }{
         err => evaluate(LettuceError(err), maxSteps - 1)(sc)
@@ -31,7 +68,11 @@ class Interpreter(private var evalConditions: EvalConditions) {
     }
   }
 
+
   override def toString: String = {
     s"INTERPRETER{ $evalConditions }"
   }
+
+
 }
+
