+++ scala-backend/app/saladbar/src/main/scala/InterpreterError.scala
@@ -1,0 +1,13 @@
+package saladbar
+
+
+/**
+  * InterpreterError
+  * 
+  * used by Interpreter class
+  *
+  * @param msg
+  */
+case class InterpreterError(msg: String) extends Exception{
+  override def toString: String = s"INTERPRETER ERROR: $msg"
+}
