+++ scala-backend/app/saladbar/src/main/scala/Main.scala
@@ -1,5 +1,11 @@
 package saladbar
 
+
+/**
+  * Main
+  * 
+  * not really used as this is a library, but here if needed for dirty non unit-testing
+  */
 object Main extends App {
   println(Let("y", N(1), N(2)))
 }
