+++ scala-backend/app/saladbar/src/main/scala/N.scala
@@ -1,0 +1,17 @@
+package saladbar
+
+
+/**
+  * N(n)
+  * concrete syntax for a number n
+  *
+  * @param n
+  */
+case class N(n: Double) extends Value {
+    def toNum = n
+    def toBool: Boolean = n != 0
+    override def toString: String = s"$n"
+    def substitute[A](evalConditions: EvalConditions, x: String, esub: Expr)(sc: Expr => A): A = 
+            sc(this)
+}
+
