+++ scala-backend/app/saladbar/src/main/scala/S.scala
@@ -1,0 +1,21 @@
+package saladbar
+
+
+/**
+  * S(s)
+  * concrete syntax of string s
+  *
+  * @param s
+  */
+case class S(s: String) extends Value {
+    def toNum = {
+        try {
+            s.toDouble
+        } catch {
+            case _: Throwable => Double.NaN
+        }
+    }
+    def toBool: Boolean = s != ""
+    override def toString: String = s"'$s'"
+    def substitute[A](evalConditions: EvalConditions, x: String, esub: Expr)(sc: Expr => A): A = sc(this) 
+}
