+++ scala-backend/app/saladbar/src/main/scala/Uop.scala
@@ -5,6 +5,7 @@ package saladbar
   * Uop
   * 
   * a collection of unary operations
+  * Completed in a single file due to size of the structures
   */
 sealed trait Uop {
     override def toString: String
